fixed prompts about format encoding options
Took 24 minutes
